vm: handle MAP_SHARED

+++ vm/types.h
@@ -18,7 +18,7 @@
 
 #include "hal/types.h"
 
-typedef u8 vm_flags_t;
+typedef u16 vm_flags_t;
 
 typedef u32 vm_attr_t;
